requirements.txt
agentscope==0.1.4
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.1.0
bleach==6.2.0
blinker==1.9.0
bm25s==0.1.10
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
dashscope==1.23.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
exceptiongroup==1.3.0
executing==2.2.0
expiringdict==1.2.2
fastapi==0.115.12
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
flake8==7.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.2.2
grpcio==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.60.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
identify==2.6.10
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inputimeout==1.0.4
ipykernel==6.29.5
ipython==8.36.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
litellm==1.69.3
llama-index==0.10.30
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-retrievers-bm25==0.2.0
llama-parse==0.4.9
llamaindex-py-client==0.1.19
loguru==0.6.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.9.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
modelscope_studio==0.0.5
multidict==6.4.3
mypy_extensions==1.1.0
myst-parser==4.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.4.8
openai==1.75.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
pre_commit==4.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.0
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.13.0
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==4.3.1
PyStemmer==2.2.0.3
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.0
rsa==4.9.1
ruff==0.11.10
safetensors==0.5.3
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinx-autobuild==2024.10.3
sphinx-gallery==0.19.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==1.0.0
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
striprtf==0.0.26
tenacity==8.5.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.12.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zhipuai==2.1.5.20250421
zipp==3.21.0
